# docs/openapi.json
{
  "info": {
    "description": "Owner-only identity management and future authenticated business API.",
    "title": "eruoo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "servers": [
    {
      "description": "Production",
      "url": "https://auth.eruoo.me"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ownerSession": {
        "description": "Better Auth owner Session cookie.",
        "in": "cookie",
        "name": "__Secure-eruoo.session_token",
        "type": "apiKey"
      },
      "apiKey": {
        "description": "Finite Better Auth API key with route-scoped permissions.",
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            }
          },
          "requestId": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "format": "uri",
            "pattern": "^https:\\/\\/auth\\.eruoo\\.me\\/problems\\/[a-z0-9]+(?:-[a-z0-9]+)*$",
            "description": "Canonical problem type URI in the https://auth.eruoo.me/problems/ namespace."
          }
        },
        "required": [
          "detail",
          "requestId",
          "status",
          "title",
          "type"
        ],
        "additionalProperties": false
      },
      "ValidationIssue": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          },
          "location": {
            "type": "string",
            "enum": [
              "body",
              "header",
              "path",
              "query"
            ]
          },
          "pointer": {
            "type": "string"
          }
        },
        "required": [
          "detail",
          "location",
          "pointer"
        ],
        "additionalProperties": false
      },
...
      "OAuthAuthorizationRevocation": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "enum": [
              "eruoo-web",
              "eruoo-desktop",
              "eruoo-mobile"
            ]
          },
          "deletedConsentCount": {
            "type": "integer",
            "minimum": 0
          },
          "revokedRefreshTokenCount": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "clientId",
          "deletedConsentCount",
          "revokedRefreshTokenCount"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/api/oauth/authorizations/{clientId}": {
      "delete": {
        "operationId": "revokeOAuthAuthorization",
        "security": [
          {
            "ownerSession": []
          }
        ],
        "tags": [
          "Security"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "clientId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "All unrevoked refresh tokens and all consents for the owner and client were revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAuthorizationRevocation"
                }
              }
            }
          },
          "400": {
            "description": "The request contains ambiguous credential carriers.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "A valid owner Session is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Recent owner authentication is required, or the client does not support owner-managed offline access.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The client or its owner authorization does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "The request body exceeds the 1 MiB operation limit.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "The owner authorization revocation rate limit was exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "The request failed unexpectedly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "The OAuth authorization database or rate limiter is unavailable, or stored state violates its invariant.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "504": {
            "description": "The request exceeded the service time limit.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}